pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["bot"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.73.0"
uvicorn = "0.17.0"
gunicorn = "20.1.0"
pandas = "1.4.0"
SQLAlchemy = "1.4.31"
PyJWT = "1.7.1"
kubernetes = "21.7.0"
psycopg2 = "2.9.3"
prometheus-client = "0.13.0"
celery = "5.2.3"
redis = "4.1.2"
"manage.py" = "0.2.10"
pyzmq = "22.3.0"
google-cloud-storage = "2.1.0"
aiosmtplib = "1.1.6"
aioboto3 = "9.6.0"
pydantic = {extras = ["dotenv"], version = "1.9.0"}
yoyo-migrations = "7.3.2"
aiopg = "1.3.3"
backoff = "^2.1.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "src.manage:server_manager"
migrate = "src.scripts.migrate:apply_migration"


[tool.black]
line-length = 90
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''